On-disk-test-cases overhaul
Some features require a particular Python version, f-strings for example are
supported only by Python 3.6+. The skipif feature of pytest did not help,
because the parametrization gets executed for all tests, regardless whether
they are skipped or not.

This use a different approch, using the pytest_generate_tests() hook: Python
snippets can express in a special header section (an initial comment with two
hashes) the required Python version, for example

  ## requires: python_version >= (3,6)

means that the test will be skipped for snakes older than version 3.6.

The header can also tell additional options the will be passed to the tested
function.

Lastly, the expected result snippet are always identified with the '.out'
suffix.